Refactor/met 1897 reduce the amount of common repository sotatest

+++ src/main/java/org/cardanofoundation/explorer/api/model/response/script/nativescript/NativeScriptResponse.java
@@ -1,12 +1,11 @@
 package org.cardanofoundation.explorer.api.model.response.script.nativescript;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 
 import lombok.*;
 
-import org.cardanofoundation.ledgersync.common.common.nativescript.ScriptType;
+import com.bloxbean.cardano.client.transaction.spec.script.ScriptType;
 
 @Getter
 @Setter
@@ -18,7 +17,7 @@ public class NativeScriptResponse {
   private Long numberOfTokens;
   private Long numberOfAssetHolders;
   private ScriptType conditionType;
-  private BigInteger required;
+  private Integer required;
   private List<String> keyHashes;
   private LocalDateTime after;
   private LocalDateTime before;
